+++ pintos/include/userprog/tss.h
@@ -6,18 +6,18 @@
 
 struct task_state {
 	uint32_t res1;
-	uint8_t rsp0;
-	uint8_t rsp1;
-	uint8_t rsp2;
-	uint8_t res2;
-	uint8_t ist1;
-	uint8_t ist2;
-	uint8_t ist3;
-	uint8_t ist4;
-	uint8_t ist5;
-	uint8_t ist6;
-	uint8_t ist7;
-	uint8_t res3;
+	uint64_t rsp0;
+	uint64_t rsp1;
+	uint64_t rsp2;
+	uint64_t res2;
+	uint64_t ist1;
+	uint64_t ist2;
+	uint64_t ist3;
+	uint64_t ist4;
+	uint64_t ist5;
+	uint64_t ist6;
+	uint64_t ist7;
+	uint64_t res3;
 	uint16_t res4;
 	uint16_t iomb;
 }__attribute__ ((packed));
